pyproject.toml
[tool.poetry]
name = "diploma_imgur"
version = "0.1.0"
description = ""
authors = ["Artyom Samoshin <artyomzhoy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
selene = "2.0.0rc4"
allure-pytest = "^2.13.2"
pytest = "^7.4.2"
jsonschema = "^4.19.1"
curlify = "^2.2.1"
python-decouple = "^3.8"


[tool.poetry.group.dev.dependencies]
pylint = "^2.17.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
